Feature gh

+++ back/src/main/java/com/developer/hostuser/service/HostUserService.java
@@ -11,6 +11,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import com.developer.email.EmailService;
 import com.developer.exception.AddException;
 import com.developer.exception.FindException;
 
@@ -19,6 +20,7 @@ import com.developer.hostuser.dto.HostUserDTO;
 import com.developer.hostuser.entity.HostUser;
 import com.developer.hostuser.repository.HostUserRepository;
 
+
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -26,6 +28,7 @@ import lombok.RequiredArgsConstructor;
 public class HostUserService {
 	
 	private final HostUserRepository hRepository;
+	private final EmailService emailService;
 	
 	ModelMapper modelMapper = new ModelMapper();
 	private Logger logger = LoggerFactory.getLogger(getClass());
@@ -138,16 +141,17 @@ public class HostUserService {
 	}
 
 	/**
-	 * 호스트 가입을 승인한다.
+	 * 호스트 가입을 승인한다. (메일 포함)
 	 * 
 	 * @author SR
 	 * @param hostId
-	 * @throws FindException
+	 * @throws FindException, Exception 
 	 */
-	public void readyOk(String hostId) throws FindException {
+	public void readyOk(String hostId) throws FindException, Exception {
 		Optional<HostUser> optHost = hRepository.findById(hostId);
 		HostUser hostEntity = optHost.get();
 		hostEntity.setReady(1);
+		emailService.hostOk(hostEntity.getEmail());
 		hRepository.save(hostEntity);
 	}
 
@@ -156,12 +160,18 @@ public class HostUserService {
 	 * 
 	 * @author SR
 	 * @param hostId
-	 * @throws RemoveException
+	 * @throws Exception 
+	 * @throws FindException 
 	 */
-	public void deleteHost(String hostId) throws RemoveException {
+	public void deleteHost(String hostId) throws FindException, Exception {
 		Optional<HostUser> optH = hRepository.findById(hostId);
-		HostUser entityH = optH.get();
-		hRepository.delete(entityH);
+		if(optH.isPresent()) {
+			HostUser entityH = optH.get();
+			emailService.hostReject(entityH.getEmail());
+			hRepository.delete(entityH);
+		} else {
+			throw new FindException("존재하지 않는 호스트 유저입니다.");
+		}
 	}
 	
 	/**
@@ -178,6 +188,7 @@ public class HostUserService {
 		if(optH.isPresent()) {
 			HostUser hostuser = optH.get();			
 			HostUserDTO.HostLoginDTO hostLoginDTO = modelMapper.map(hostuser, HostUserDTO.HostLoginDTO.class);
+			System.out.println("호스트아이디는 "+HostId+"비밀번호는"+pwd);
 			if(hostLoginDTO.getPwd().equals(pwd) && !hostLoginDTO.getReady().equals(2)) {
 				hostLoginDTO.setPwd("");
 				return hostLoginDTO;
@@ -204,5 +215,62 @@ public class HostUserService {
 		}
 		throw new FindException("아이디에 해당하는 고객이 없습니다");
 	}
+	
+	/**
+	 * 호스트 아이디 찾기
+	 * @author choigeunhyeong
+	 * @param Num 사업자 번호
+	 * @return
+	 * @throws FindException
+	 */
+	public HostUserDTO findHostId(String num) throws FindException{
+		Optional<HostUser> optH = hRepository.findByNum(num);
+		if(optH.isPresent()) {
+			HostUser host = optH.get();
+			HostUserDTO hostDTO = modelMapper.map(host, HostUserDTO.class);
+			if(hostDTO.getNum().equals(num)) {
+				return hostDTO;
+			}
+		}
+		throw new FindException("사업자번호가 일치하는 회원이 없습니다.");
+	}
 
+	/**
+	 * Email을 통해 해당 email로 가입된 정보가 있는지 확인. 
+	 * 가입된 정보가 있다면 입력받은 id와 email이 서로 일치한지 여부를 리턴하면서 임시비밀번호로 변경 및 메일발송
+	 * @param email
+	 * @param hostId
+	 * @return
+	 * @throws Exception
+	 */
+	public boolean hostPwdAndEmailCheck(String email, String hostId) throws Exception {
+		HostUser host = hRepository.hostEmailCheck(email);
+        if(host!=null && host.getHostId().equals(hostId)) {
+        	String temporaryPwd = emailService.updatePwd(email);
+        	host.setPwd(temporaryPwd);
+        	hRepository.save(host);
+        	return true;
+        }
+        else {
+            return false;
+        }
+	}
+	
+	
+	/**
+	 * 본인인증 이메일 체크(가입여부확인)
+	 * @author SR
+	 * @param email
+	 * @return true: 신규가입가능 false: 신규가입불가
+	 */
+	public boolean hostEmailCheck(String email){
+		HostUser host = hRepository.hostEmailCheck(email);
+        if(host==null) {
+        	return true; //가입된 정보가 없음
+        }
+        else{
+            return false; //가입된 정보가 있음
+        }
+	}
+	
 }
